make modules according t task - but it"s fucking __dirname give mistakes
