+++ .gitignore
@@ -13,5 +13,7 @@ node_modules/
 
 # Папка с собранными файлами проекта
 build/
+
+# сбилденные вебпаком файлы
 bundle.js
 bundle.js.map
