+++ package.json
@@ -5,6 +5,7 @@
   "description": "Личный проект «Менеджер задач» от HTML Academy",
   "scripts": {
     "build": "webpack",
+    "start": "webpack-dev-start --open",
     "eslint": "eslint src/**/*.js",
     "test": "npm run eslint"
   },
@@ -20,5 +21,8 @@
   },
   "engines": {
     "node": "10"
+  },
+  "dependencies": {
+    "@types/node": "^11.9.5"
   }
 }
