+++ public/index.html
@@ -145,7 +145,7 @@
       </section>
     </main>
 
-    <script src="../src/bundle.js"></script>
+    <script src="bundle.js"></script>
 
   </body>
 </html>
